Architecture closer, edit Op n View
Clock now spawns another dialog box to resize the text of the clock. It doesn't resize the clock frame though. Also, after that, the edit button flickers in & out of visibility. le sigh.

Nonetheless, this is closer to the architecture I'm intending to emulate. Some of the message passing is truncated into runclock rather than run. And I made a hard assumption for updatesettings. I'll have to split who's listening to the buttons or check how other programs have a single object listen to more than one JComponent. After I fix those, I'll add more of the settings to the ConfigPair so the clock can change arbitrarily.