Proof concept: message passing
Moving to a design where I pass custom messages rather than swing Events. This way the view maps buttons/textfields to messages rather than the model or operationObj. I could push this farther and have the operation send a message too, but I think that could push me a bit far into the 'reinventing methods via enums' territory.

I'll meditate a bit more on the two way messaging in a moment. Otherwise, I need to pull out the previous way of doing that, as RunClock is still generating all the objects. That needs to be pushed up into Run or the ViewRoot.