More config hooks
Added more spinners, fields, events, flags, and so on to change the frame w/h and clock w/h. The former seems to fight with the gridbaglayout somewhat, because it will jump to a packed-type size depending upon which way the spinner is moved. The latter seems not to work at all. To be fair, there's no padding, so I don't expect much. Perhaps the clock w/h changes should affect the layout rather than the pane. I don't know right now.

The communication path is a bit tedious to update (and prone to copypasta errors). After all these are done, I'll start drawing the fancy versions.